+++ NEWS.md
@@ -15,6 +15,8 @@ function `AnnotateIDVCF` with default value TRUE.
 `ReadAndSplitStrelkaSBSVCFs` under one element `discarded.variants` in the
 return value. An extra column `discarded.reason` were added to show the details.
 
+* Updated internal functions `ReadVCF` and `ReadVCFs` not to remove any discarded variants.
+
 ## Fixed
 * Fixed bugs in `CheckAndReturnSBSMatrix`, `CheckAndReturnDBSMatrix`, 
 `CreateOneColSBSMatrix`,`CreateOneColDBSMatrix`, `VCFsToSBSCatalogs`,